drivers/requirements.txt
pyrtlsdr
requests
websocket-client